requirements.txt
absl-py
atari-py
attrs
dopamine-rl
gin-config
gym==0.19.0
matplotlib
monty
more_itertools
networkx
numpy
pandas
recsim
seaborn
simplejson
sklearn
statsmodels
tensorflow
tqdm
